requirements.txt
tensorflow==2.13.0
numpy==1.24.3
opencv-python==4.8.0.76
albumentations==1.3.1
nibabel==5.1.0
scikit-image==0.21.0
matplotlib==3.7.2
pyyaml==6.0.1
tqdm==4.65.0
